@@ -20,66 +20,55 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "9.1.4",
-        "@angular/cdk": "9.2.2",
-        "@angular/common": "9.1.4",
-        "@angular/compiler": "9.1.4",
-        "@angular/core": "9.1.4",
-        "@angular/forms": "9.1.4",
-        "@angular/material": "9.2.2",
-        "@angular/material-moment-adapter": "9.2.2",
-        "@angular/platform-browser": "9.1.4",
-        "@angular/platform-browser-dynamic": "9.1.4",
-        "@angular/router": "9.1.4",
-        "@fullcalendar/angular": "4.4.5-beta",
-        "@fullcalendar/core": "4.4.0",
-        "@fullcalendar/daygrid": "4.4.0",
-        "@fullcalendar/interaction": "4.4.0",
-        "@fullcalendar/list": "4.4.0",
-        "@fullcalendar/moment": "4.4.0",
-        "@fullcalendar/rrule": "4.4.0",
-        "@fullcalendar/timegrid": "4.4.0",
-        "@types/humanize-duration": "^3.18.1",
-        "apexcharts": "3.19.2",
-        "crypto-js": "3.3.0",
-        "highlight.js": "10.0.1",
-        "humanize-duration": "^3.24.0",
-        "lodash": "4.17.15",
-        "moment": "2.24.0",
-        "ng-apexcharts": "1.5.12",
-        "ngx-markdown": "9.0.0",
-        "ngx-quill": "9.1.0",
-        "perfect-scrollbar": "1.5.0",
-        "quill": "1.3.7",
-        "rrule": "2.6.4",
-        "rxjs": "6.5.5",
-        "tslib": "1.11.1",
-        "web-animations-js": "2.3.2",
-        "zone.js": "0.10.3"
+        "@angular/animations": "v13-lts",
+        "@angular/cdk": "v13-lts",
+        "@angular/common": "v13-lts",
+        "@angular/compiler": "v13-lts",
+        "@angular/core": "v13-lts",
+        "@angular/forms": "v13-lts",
+        "@angular/material": "v13-lts",
+        "@angular/material-moment-adapter": "v13-lts",
+        "@angular/platform-browser": "v13-lts",
+        "@angular/platform-browser-dynamic": "v13-lts",
+        "@angular/router": "v13-lts",
+        "@types/humanize-duration": "^3.27.1",
+        "crypto-js": "^4.1.1",
+        "highlight.js": "^11.6.0",
+        "humanize-duration": "^3.27.3",
+        "lodash": "4.17.21",
+        "moment": "^2.29.4",
+        "ng-apexcharts": "^1.7.4",
+        "ngx-markdown": "^13.1.0",
+        "perfect-scrollbar": "^1.5.5",
+        "quill": "^1.3.7",
+        "rrule": "^2.7.1",
+        "rxjs": "^7.5.7",
+        "tslib": "^2.4.1",
+        "web-animations-js": "^2.3.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "0.901.4",
-        "@angular/cli": "9.1.4",
-        "@angular/compiler-cli": "9.1.4",
-        "@angular/language-service": "9.1.4",
-        "@types/crypto-js": "3.1.45",
-        "@types/highlight.js": "9.12.3",
-        "@types/jasmine": "3.5.10",
-        "@types/jasminewd2": "2.0.8",
-        "@types/lodash": "4.14.150",
-        "@types/node": "12.12.37",
-        "codelyzer": "5.2.2",
-        "jasmine-core": "3.5.0",
-        "jasmine-spec-reporter": "4.2.1",
-        "karma": "5.0.4",
-        "karma-chrome-launcher": "3.1.0",
-        "karma-coverage-istanbul-reporter": "2.1.1",
-        "karma-jasmine": "3.0.3",
-        "karma-jasmine-html-reporter": "1.5.3",
-        "protractor": "5.4.4",
-        "tailwindcss": "1.4.4",
-        "ts-node": "8.3.0",
-        "tslint": "6.1.2",
-        "typescript": "3.8.3"
+        "@angular-devkit/build-angular": "v13-lts",
+        "@angular/cli": "v13-lts",
+        "@angular/compiler-cli": "v13-lts",
+        "@angular/language-service": "v13-lts",
+        "@types/crypto-js": "^4.1.1",
+        "@types/highlight.js": "^10.1.0",
+        "@types/jasmine": "^4.3.0",
+        "@types/jasminewd2": "^2.0.10",
+        "@types/lodash": "^4.14.188",
+        "@types/node": "^18.11.9",
+        "codelyzer": "^6.0.2",
+        "jasmine-core": "^4.5.0",
+        "jasmine-spec-reporter": "^7.0.0",
+        "karma": "^6.4.1",
+        "karma-chrome-launcher": "^3.1.1",
+        "karma-coverage": "^2.2.0",
+        "karma-jasmine": "^5.1.0",
+        "karma-jasmine-html-reporter": "^2.0.0",
+        "protractor": "^7.0.0",
+        "tailwindcss": "^3.2.3",
+        "ts-node": "^10.9.1",
+        "tslint": "^6.1.3",
+        "typescript": "^4.6.4"
     }
 }
